Old dependencies cleared. Concordance is now operational, but limited. Bugs in TextService.java logic has been removed. Script for graphics generation has been fixed.

+++ src/main/java/org/edloidas/text/words/content/Score.java
@@ -88,9 +88,9 @@ public class Score {
      */
     public void calculate(int totalCount, int neighborCount, int wordsCount) {
         try {
-            int S = 2;
-            double p = totalCount / (wordsCount - neighborCount);
-            double E = p * neighborCount * S;
+            double S = 2.0d;
+            double p = (double) totalCount / (double) (wordsCount - neighborCount);
+            double E = p * (double) neighborCount * S;
             double q = 1.0d - p;
             this.score =  Math.pow(Math.abs(((double)this.count - E)), (E * q));
         } catch (Exception ex) {
